+++ tradingagents/graph/__init__.py
@@ -2,6 +2,7 @@
 
 from .trading_graph import TradingAgentsGraph
 from .conditional_logic import ConditionalLogic
+from .node_timing import NodeExecutionRecord, NodeTimingTracker
 from .setup import GraphSetup
 from .propagation import Propagator
 from .reflection import Reflector
@@ -10,6 +11,8 @@ from .signal_processing import SignalProcessor
 __all__ = [
     "TradingAgentsGraph",
     "ConditionalLogic",
+    "NodeExecutionRecord",
+    "NodeTimingTracker",
     "GraphSetup",
     "Propagator",
     "Reflector",